adminclear

+++ app.py
@@ -290,7 +290,8 @@ def login():
 @app.route("/adminclear")
 def adminclear():
     for f in os.listdir(UPLOAD_FOLDER):
-        os.remove(os.path.join(UPLOAD_FOLDER, f))
+        if f != "keep.txt":
+            os.remove(os.path.join(UPLOAD_FOLDER, f))
     return redirect(url_for("mainpage"))
 
 @app.route('/delete/<int:id>')
